backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-jwt-extended
python-dotenv
werkzeug
tensorflow
numpy
pillow
opencv-python-headless
reportlab
gunicorn
